feature(event-section): rename CarouselItem to EventCarouselItem
