+++ client/src/components/EventCarouselItem/index.jsx
