+++ client/src/components/EventCarouselItem/styles.js
